move valence utils to packages
