+++ packages/Cargo.toml
@@ -8,3 +8,7 @@ alloy                               = { workspace = true }
 serde                               = { workspace = true }
 valence-clearing-queue-supervaults  = { workspace = true }
 log                                 = { workspace = true }
+valence-domain-clients              = { workspace = true }
+valence-authorization-utils         = { workspace = true }
+cosmwasm-std                        = { workspace = true }
+valence-processor-utils             = { workspace = true }
