+++ strategies/wbtc_test/strategist/src/lib.rs
@@ -1,4 +1,3 @@
 pub mod phases;
 pub mod strategist;
 pub mod strategy_config;
-pub mod valence_utils;
